% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisions.R
\name{mvrs_decision_node}
\alias{mvrs_decision_node}
\title{Create a single decision node}
\usage{
mvrs_decision_node(name, branches)
}
\arguments{
\item{name}{Character string naming this decision.}

\item{branches}{Character vector of possible choices.}
}
\value{
An S3 object of class \code{mvrs_decision_node}.
}
\description{
A decision node represents one analyst degree of freedom with multiple
possible branches (choices).
}
