% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisions.R
\name{mvrs_decisions}
\alias{mvrs_decisions}
\title{Define the multiverse decision space}
\usage{
mvrs_decisions(...)
}
\arguments{
\item{...}{\code{mvrs_decision_node} objects.}
}
\value{
An S3 object of class \code{mvrs_decisions}.
}
\description{
Manually specifies the set of decision nodes and their branches that
define the multiverse of analyses.
}
