% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{mvrs_results}
\alias{mvrs_results}
\title{Create a multiverse results object}
\usage{
mvrs_results(specs, estimates, question, dataset_name)
}
\arguments{
\item{specs}{A tibble of specification combinations.}

\item{estimates}{A tibble of results with columns \code{estimate}, \code{std_error},
\code{p_value}, and \code{significant}.}

\item{question}{The \code{mvrs_question} used.}

\item{dataset_name}{Name of the dataset analysed.}
}
\value{
An S3 object of class \code{mvrs_results}.
}
\description{
Create a multiverse results object
}
\keyword{internal}
