% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{mvrs_spec_curve}
\alias{mvrs_spec_curve}
\title{Specification curve plot}
\usage{
mvrs_spec_curve(results)
}
\arguments{
\item{results}{A \code{mvrs_results} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Produces a specification curve visualization: a panel of sorted effect
estimates with indicators for which decisions were active in each
specification.
}
